+++ thomas/ThomasCore/src/thomas/graphics/animation/AnimPlayback.cpp
@@ -29,7 +29,7 @@ namespace thomas {
 			{
 				m_channel.reserve(anim.numChannel());
 				//Each channel needs to be initiated so it can generate it's own frames.
-				for (size_t i = 0; i < anim.numChannel(); i++) {
+				for (uint32_t i = 0; i < anim.numChannel(); i++) {
 					const ObjectChannel& ch = anim[i];
 					unsigned int boneInd;
 					if(ch.numNodeChannels() == 0 ||
